base/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
Flask==3.1.0
GDAL @ file:///C:/Users/VISHWAS/Downloads/GDAL-3.6.1-cp310-cp310-win_amd64.whl#sha256=30340137fea7c080f294ea0b93770a892e7b1c68ae5f9c8a87b6b59c78eff0c9
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Js2Py==0.74
lxml==5.3.0
MarkupSafe==3.0.2
packaging==24.2
phonenumbers==8.13.49
pillow==11.0.0
pipwin==0.5.2
psycopg2==2.9.10
pyjsparser==2.7.1
PyJWT==2.9.0
PyPrind==2.11.3
pySmartDL==1.3.4
python-decouple==3.8
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
